respect and synchronize with user prefs
