+++ src/gwt/panmirror/src/editor/src/optional/ace/ace_editor.ts
@@ -21,6 +21,7 @@ export interface ChunkEditorFactory {
 
 export interface AceChunkEditor {
     editor: AceAjax.Editor;
-    setMode (mode: string): void;
+    setMode(mode: string): void;
     element: HTMLElement;
+    destroy(): void;
 }
