+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTarget.java
@@ -1567,7 +1567,8 @@ public class TextEditingTarget implements
          }.schedule(100);
       }
 
-      registerPrefs(releaseOnDismiss_, prefs_, projConfig_, docDisplay_, document);
+      TextEditingTargetPrefsHelper.registerPrefs(
+            releaseOnDismiss_, prefs_, projConfig_, docDisplay_, document);
 
       // Initialize sourceOnSave, and keep it in sync. Don't source on save
       // (regardless of preference) in auto save mode, which is mutually
@@ -7015,7 +7016,7 @@ public class TextEditingTarget implements
       return docUpdateSentinel_.getPath() == null;
    }
 
-   private static boolean shouldEnforceHardTabs(FileSystemItem item)
+   public static boolean shouldEnforceHardTabs(FileSystemItem item)
    {
       if (item == null)
          return false;
@@ -7113,198 +7114,6 @@ public class TextEditingTarget implements
    public CompletionContext getRCompletionContext()
    {
       return rContext_;
-   }
-
-   // these methods are public static so that other editing targets which
-   // display source code (but don't inherit from TextEditingTarget) can share
-   // their implementation
-
-   public static interface PrefsContext
-   {
-      FileSystemItem getActiveFile();
-   }
-
-   public static void registerPrefs(
-                     ArrayList<HandlerRegistration> releaseOnDismiss,
-                     UserPrefs prefs,
-                     ProjectConfig projectConfig,
-                     DocDisplay docDisplay,
-                     final SourceDocument sourceDoc)
-   {
-      registerPrefs(releaseOnDismiss,
-                    prefs,
-                    projectConfig,
-                    docDisplay,
-                    new PrefsContext() {
-                        @Override
-                        public FileSystemItem getActiveFile()
-                        {
-                           String path = sourceDoc.getPath();
-                           if (path != null)
-                              return FileSystemItem.createFile(path);
-                           else
-                              return null;
-                        }
-                    });
-   }
-
-   public static void registerPrefs(
-                     ArrayList<HandlerRegistration> releaseOnDismiss,
-                     UserPrefs prefs,
-                     final ProjectConfig projectConfig,
-                     final DocDisplay docDisplay,
-                     final PrefsContext context)
-   {
-      releaseOnDismiss.add(prefs.highlightSelectedLine().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setHighlightSelectedLine(arg);
-               }}));
-      releaseOnDismiss.add(prefs.highlightSelectedWord().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setHighlightSelectedWord(arg);
-               }}));
-      releaseOnDismiss.add(prefs.showLineNumbers().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setShowLineNumbers(arg);
-               }}));
-      releaseOnDismiss.add(prefs.useSpacesForTab().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  if (shouldEnforceHardTabs(context.getActiveFile()))
-                  {
-                     docDisplay.setUseSoftTabs(false);
-                  }
-                  else
-                  {
-                     if (projectConfig == null)
-                        docDisplay.setUseSoftTabs(arg);
-                  }
-               }}));
-      releaseOnDismiss.add(prefs.numSpacesForTab().bind(
-            new CommandWithArg<Integer>() {
-               public void execute(Integer arg) {
-                  if (projectConfig == null)
-                     docDisplay.setTabSize(arg);
-               }}));
-      releaseOnDismiss.add(prefs.autoDetectIndentation().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  if (projectConfig == null)
-                     docDisplay.autoDetectIndentation(arg);
-               }}));
-      releaseOnDismiss.add(prefs.showMargin().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setShowPrintMargin(arg);
-               }}));
-      releaseOnDismiss.add(prefs.blinkingCursor().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setBlinkingCursor(arg);
-               }}));
-      releaseOnDismiss.add(prefs.marginColumn().bind(
-            new CommandWithArg<Integer>() {
-               public void execute(Integer arg) {
-                  docDisplay.setPrintMarginColumn(arg);
-               }}));
-      releaseOnDismiss.add(prefs.showInvisibles().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setShowInvisibles(arg);
-               }}));
-      releaseOnDismiss.add(prefs.showIndentGuides().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setShowIndentGuides(arg);
-               }}));
-      releaseOnDismiss.add(prefs.scrollPastEndOfDocument().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setScrollPastEndOfDocument(arg);
-               }}));
-      releaseOnDismiss.add(prefs.highlightRFunctionCalls().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setHighlightRFunctionCalls(arg);
-               }}));
-      releaseOnDismiss.add(prefs.rainbowParentheses().bind(
-          new CommandWithArg<Boolean>() {
-             public void execute(Boolean arg) {
-                docDisplay.setRainbowParentheses(arg);
-             }}));
-      releaseOnDismiss.add(prefs.editorKeybindings().bind(
-            new CommandWithArg<String>() {
-               public void execute(String arg) {
-                  docDisplay.setUseVimMode(arg == UserPrefs.EDITOR_KEYBINDINGS_VIM);
-               }}));
-      releaseOnDismiss.add(prefs.editorKeybindings().bind(
-            new CommandWithArg<String>() {
-               public void execute(String arg) {
-                  docDisplay.setUseEmacsKeybindings(arg == UserPrefs.EDITOR_KEYBINDINGS_EMACS);
-               }}));
-      releaseOnDismiss.add(prefs.codeCompletionOther().bind(
-            new CommandWithArg<String>() {
-               public void execute(String arg) {
-                  docDisplay.syncCompletionPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.codeCompletionCharacters().bind(
-            new CommandWithArg<Integer>() {
-               public void execute(Integer arg) {
-                  docDisplay.syncCompletionPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.codeCompletionDelay().bind(
-            new CommandWithArg<Integer>() {
-               public void execute(Integer arg) {
-                  docDisplay.syncCompletionPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.enableSnippets().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.syncCompletionPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.showDiagnosticsOther().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.syncDiagnosticsPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.diagnosticsOnSave().bind(
-            new CommandWithArg<Boolean>() {
-               @Override
-               public void execute(Boolean arg)
-               {
-                  docDisplay.syncDiagnosticsPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.backgroundDiagnosticsDelayMs().bind(
-            new CommandWithArg<Integer>() {
-               public void execute(Integer arg) {
-                  docDisplay.syncDiagnosticsPrefs();
-               }}));
-      releaseOnDismiss.add(prefs.showInlineToolbarForRCodeChunks().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.forceImmediateRender();
-               }}));
-      releaseOnDismiss.add(prefs.foldStyle().bind(
-            new CommandWithArg<String>() {
-               public void execute(String style)
-               {
-                  docDisplay.setFoldStyle(FoldStyle.fromPref(style));
-               }}));
-      releaseOnDismiss.add(prefs.surroundSelection().bind(
-            new CommandWithArg<String>() {
-               public void execute(String string)
-               {
-                  docDisplay.setSurroundSelectionPref(string);
-               }}));
-      releaseOnDismiss.add(prefs.enableTextDrag().bind(
-            new CommandWithArg<Boolean>() {
-               public void execute(Boolean arg) {
-                  docDisplay.setDragEnabled(arg);
-               }}));
-
    }
 
    public static void syncFontSize(
